minor view refactorings
