+++ packages/view/src/drivers/handlebars/helpers/json.ts
