+++ packages/view/src/drivers/handlebars/helpers/stack.ts
