+++ packages/view/src/drivers/handlebars/index.ts
@@ -1,0 +1,2 @@
+
+export * from './handlebars-driver.js'
