--- packages/view/src/engines/handlebars/index.ts
@@ -1,2 +1,0 @@
-
-export * from './compiler.js'
