+++ packages/view/src/index.ts
@@ -1,5 +1,5 @@
 
-export { HandlebarsCompiler } from './engines/index.js'
+export { HandlebarsDriver } from './drivers/index.js'
 export { ViewConfigBuilder } from './view-config-builder.js'
 export { ViewManager } from './view-manager.js'
 export { ViewResponse } from './view-response.js'
